[ZPI-73] - Refactor structre
